Use int64 type

+++ github-repo-crawler/models/commit.go
@@ -74,7 +74,7 @@ func (m *CommitDB) GetLastCommitByRepoID(ctx context.Context, repoID int) (*Comm
 	return &native, err
 }
 
-func (m *CommitDB) GetFirstCommitByRepoID(ctx context.Context, repoID int) (*Commit, error) {
+func (m *CommitDB) GetFirstCommitByRepoID(ctx context.Context, repoID int64) (*Commit, error) {
 
 	var native Commit
 	err := m.Db.Table(m.TableName()).Where("repository_id = ?", repoID).Order("created_at ASC LIMIT 1").Find(&native).Error
